warehouse viewer (WIP)

+++ src/main/java/KursovProektOOP2/data/repository/AgentRepository.java
@@ -64,7 +64,7 @@ public class AgentRepository implements DAORepository {
     }
 
     @Override
-    public Optional getById(Long id) {
+    public Optional getById(int id) {
         Session session = Connection.openSession();
         Transaction transaction = session.beginTransaction();
         Optional<Agent> foundAgent = null;
