+++ src/main/java/KursovProektOOP2/data/repository/RoleRepository.java
@@ -67,7 +67,7 @@ public class RoleRepository implements DAORepository {
     }
 
     @Override
-    public Optional getById(Long id) {
+    public Optional getById(int id) {
         Session session=Connection.openSession();
         Transaction transaction=session.beginTransaction();
         Optional<Role> foundRole=null;
