+++ src/main/java/KursovProektOOP2/data/repository/UserNotificationRepository.java
@@ -65,7 +65,7 @@ public class UserNotificationRepository implements DAORepository {
     }
 
     @Override
-    public Optional getById(Long id) {
+    public Optional getById(int id) {
         Session session=Connection.openSession();
         Transaction transaction=session.beginTransaction();
         Optional<Usernotifications> foundUserNotification=null;
